feat(AHC033): Use one small crane
FIXME: if crane is forced to drop container to the wrong space,
this code panics. This issue occurs in about 1% of test cases.

This commit I could NOT take care of legacy comments. Me tomorrow
could handle this, I hope.

Sum of scores of 100 sample cases: 21,904 (changes because of rand)